Refactoring effectors our of background
