+++ Editor/app.js
@@ -2,8 +2,8 @@ import { scenes } from './data/scenes.js';
 import { arpObjects } from "./data/arpObjects.js";
 import { start, stop, addArp, deleteArp, adjustVolume, nextSynth } from "./audio/arpeggiator.js";
 import { moods } from "./data/moods-chords.js";
-import { initEffector, setEffect } from "./audio/effector.js";
-import { setBackround, updateEffectorVisibility } from './utils/background.js';
+import { initEffectors, updateEffectorVisibility } from "./audio/effector.js";
+import { setBackround,  } from './utils/background.js';
 import { toggleDebugMode } from './settings.js';
 
 //event lisener for page load
