+++ Editor/audio/effector.js
@@ -1,47 +1,134 @@
-import { getSprite, getSprites } from "../app.js";
-import { getArp } from "./arpeggiator.js";
 
-//TODO: A map (or other data structure) containing the effect and wetness for the target sprites
-let effectWetness = 0;
-let bitCrusherSprite = undefined;
-let sprite = undefined;
-let effect = undefined;
-let fragmentShader = undefined;
+import { settings } from '../settings.js';
+import { getArp } from '../audio/arpeggiator.js';
+import { bloomShader } from '../shaders/bloom.js';
+import { glitchShader } from '../shaders/glitch.js';
+import { noiseShader } from '../shaders/noise.js';
+import { pixelateShader }   from '../shaders/pixelate.js';
+import { rippleShader } from '../shaders/ripple.js';
+import { trailsShader } from '../shaders/trails.js';
+import { waveShader } from '../shaders/wave.js';
 
-//TODO: this needs to be generic
-let effectFilter = new PIXI.Filter(null, fragmentShader, {
-    pixelSize: .001, // Initial pixel size
-});
+//#region Effectors
+let effectors = [];
 
-//TODO: review...
-export function setEffect(spriteId) {
-    let arp = getArp(spriteId);   //affect the audio
-    sprite = getSprite(spriteId); //affect the visual
+export function initEffectors(app, transport, sceneName) {
+    const effectorDefinitions = scenes[sceneName].effectors;
+    effectors.forEach(rect => app.stage.removeChild(rect)); // Clear existing rectangles
+    effectors = [];
 
-    //TODO: provide an option for simple overlap or distance based activation
-    if(distance(sprite.x, sprite.y, bitCrusherSprite.x, bitCrusherSprite.y) < 200) { 
-        crusher.wet.value = effectWetness;
-        arp.synth.connect(crusher);
-        sprite.filters = [effectFilter];
-        sprite.effectActive = true;
+    //Load the effects, shaders and rectangle
+    effectorDefinitions.forEach((effectorDefinition) => {
+        const rectangle = new PIXI.Graphics();
+        rectangle.beginFill(0x66CCFF);
+        rectangle.x = effectorDefinition.position[0];
+        rectangle.y = effectorDefinition.position[1];
+        rectangle.width = effectorDefinition.size[0];
+        rectangle.height = effectorDefinition.size[1];
+        rectangle.drawRect(0, 0, effectorDefinition.size[0], effectorDefinition.size[1]);
+        rectangle.alpha = settings.debugMode ? 0.5 : 0;
+
+        const effectText = new PIXI.Text(effectorDefinition.effect, {
+            fontFamily: 'Arial',
+            fontSize: 14,
+            fill: 0xffffff,
+            align: 'center'
+        });
+        
+        // Center the text on the sprite
+        effectText.anchor.set(0.5);
+        effectText.x = rectangle.width / 2;
+        effectText.y = rectangle.height / 2;
+            
+        rectangle.addChild(effectText);
+
+        app.stage.addChild(rectangle);
+
+        var effect = lookUpEffect(effectorDefinition.effect);
+
+        //load the shader file
+        let shaderFile = lookUpShader(effectorDefinition.shader);
+        let effectFilter = new PIXI.Filter(null, shaderFile, {
+            pixelSize: .001, // TODO: use a common parameter for all shaders
+        });
+
+        effectors.push({rectangle, effect, effectFilter});
+    });
+
+    //TODO: Upon adding a sprite to the scene, add all shaders and effects.  Then control intensity based on distance from affector rectangle
+
+    //PIXI update loop
+    app.ticker.add((delta) => {
+        effectors.forEach(eff => {
+            const sprites = app.stage.children.filter(child => child instanceof PIXI.Sprite);
+            sprites.forEach(sprite => {
+                
+                //If the sprite is within the rectangle - TODO: distance based falloff
+                if (sprite.x > eff.rectangle.x && sprite.x < eff.rectangle.x + eff.rectangle.width &&
+                    sprite.y > eff.rectangle.y && sprite.y < eff.rectangle.y + eff.rectangle.height) {
+                    // Apply the effect
+                    console.log("Setting " + eff.effect + " on " + sprite.id);
+                    eff.rectangle.tint = 0xFF0000;
+                    const arp = getArp(sprite.id);
+                    arp.synth.connect(eff.effect);
+                    sprite.filters = [eff.effectFilter];
+                } else {
+                    // Remove the effect
+                    eff.rectangle.tint = 0xFFFFFF;
+                    const arp = getArp(sprite.id);
+                    //arp.synth.disconnect(eff.effect);
+                    //sprite.filters = [];
+                }
+            });
+        });
+    });
+}
+
+export function updateEffectorVisibility() {
+    effectors.forEach(eff => {
+        eff.rectangle.alpha = settings.debugMode ? 0.5 : 0;
+    });
+}
+//#endregion
+
+//#region Utilities
+//TODO: Do we need to control the initial parameters
+function lookUpEffect(effectName) {
+    switch (effectName) {
+        
+        case "chorus":
+            return new Tone.Chorus(4, 2.5, 0.5); 
+        case "bitCrusher":
+            return new Tone.BitCrusher(4);
+        case "reverb":
+            return new Tone.Reverb(1.5);
+        case "delay":
+            return new Tone.FeedbackDelay("8n", 0.5);
+        case "distortion":
+            return new Tone.Distortion(0.4);
+        case "phaser":
+            return new Tone.Phaser(15, 5, 1000);
+        case "tremolo":
+            return new Tone.Tremolo(9, 0.75);
     }
-    else {
-        arp.synth.disconnect(crusher);
-        sprite.filters = [];
-        sprite.effectActive = false;
+}
+
+function lookUpShader(effectorName) {
+    switch (effectorName) {
+        case "bloomShader":
+            return bloomShader;
+        case "glitchShader":
+            return glitchShader;
+        case "noiseShader":
+            return noiseShader;
+        case "pixelateShader":
+            return pixelateShader;
+        case "rippleShader":
+            return rippleShader;
+        case "trailsShader":
+            return trailsShader;
+        case "waveShader":
+            return waveShader;
     }
 }
-
-export function initEffector(app, effect, shader, position, size) {
-    //load the shader from the shader file
-    
-
-}
-
-export function updateEffector(delta) {
-    console.log("Updating effector");
-}
-
-function distance(x1, y1, x2, y2) {
-    return Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2));
-  }
+//#endregion
